+++ ep1/ep1_ciclista.h
@@ -1,6 +1,6 @@
 #ifndef CICLISTA_H
 #define CICLISTA_H
 
-int cur_ticket;
-int fetch_and_add(int *variable, int value);
+unsigned long int cur_ticket;
+int fetch_and_add(unsigned long int *variable, unsigned long int value);
 #endif
